backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pydantic
sqlalchemy
python-multipart

